Time: 4 ms (56.43%), Space: 5.9 MB (25.42%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -3,7 +3,7 @@ public:
     int mySqrt(int x) {
         int start = 1, end = x, ans = 0;
         while(start <= end) {
-            long long int mid = start + (end - start) / 2;
+            long mid = start + (end - start) / 2;
             if(mid * mid == x) { return mid; }
             else if(mid * mid < x) {
                 ans = mid;
